Recipes

AlgebraOfGraphics.choropleth! — Function

choropleth(geometries; transformation, attributes...)

Choropleth map, where regions are defined by geometries. Use transformation to transform coordinates (see Proj.jl for more information).

Warning

The transformation keyword argument is experimental and could be deprecated (even in a non-breaking release) in favor of a different syntax.

Attributes

Available attributes and their defaults for Combined{AlgebraOfGraphics.choropleth!} are:

AlgebraOfGraphics.choropleth — Function

choropleth(geometries; transformation, attributes...)

Choropleth map, where regions are defined by geometries. Use transformation to transform coordinates (see Proj.jl for more information).

Warning

The transformation keyword argument is experimental and could be deprecated (even in a non-breaking release) in favor of a different syntax.

Attributes

Available attributes and their defaults for Combined{AlgebraOfGraphics.choropleth} are:

  color         :gray25
  colormap      :batlow
  colorrange    MakieCore.Automatic()
  cycle         [:color => :patchcolor]
  highclip      MakieCore.Automatic()
  inspectable   true
  linestyle     "nothing"
  lowclip       MakieCore.Automatic()
  nan_color     :transparent
  overdraw      false
  shading       false
  space         :data
  strokecolor   :black
  strokewidth   0
  transparency  false
  visible       true

AlgebraOfGraphics.linesfill! — Function

linesfill(xs, ys; lower, upper, attributes...)

Line plot with a shaded area between lower and upper. If lower and upper are not given, shaded area is between 0 and ys.

Attributes

Available attributes and their defaults for Combined{AlgebraOfGraphics.linesfill!} are:

AlgebraOfGraphics.linesfill — Function

linesfill(xs, ys; lower, upper, attributes...)

Line plot with a shaded area between lower and upper. If lower and upper are not given, shaded area is between 0 and ys.

Attributes

Available attributes and their defaults for Combined{AlgebraOfGraphics.linesfill} are:

  color       :gray25
  colormap    :batlow
  colorrange  MakieCore.Automatic()
  fillalpha   0.15
  linestyle   "nothing"
  linewidth   1.5
  lower       MakieCore.Automatic()
  upper       MakieCore.Automatic()