Types

AlgebraOfGraphics.AbstractDrawable — Type

AbstractDrawable

Abstract type encoding objects that can be drawn via AlgebraOfGraphics.draw.

AlgebraOfGraphics.AbstractAlgebraic — Type

AbstractAlgebraic  <: AbstractDrawable

Abstract type encoding objects that can be combined together using + and *.

AlgebraOfGraphics.Layer — Type

Layer(transformation, data, positional::AbstractVector, named::AbstractDictionary)

Algebraic object encoding a single layer of a visualization. It is composed of a dataset, positional and named arguments, as well as a transformation to be applied to those. Layer objects can be multiplied, yielding a novel Layer object, or added, yielding a AlgebraOfGraphics.Layers object.

AlgebraOfGraphics.Layers — Type

Layers(layers::Vector{Layer})

Algebraic object encoding a list of AlgebraOfGraphics.Layer objects. Layers objects can be added or multiplied, yielding a novel Layers object.

AlgebraOfGraphics.ProcessedLayer — Type

ProcessedLayer(l::Layer)

Process a Layer and return the resulting ProcessedLayer.

Note that this method should not be used anymore as processing a Layer can now potentially return multiple ProcessedLayer objects. Therefore, you should use the plural form ProcessedLayers(layer).

AlgebraOfGraphics.ProcessedLayers — Type

ProcessedLayers(layers::Vector{ProcessedLayer})

Object encoding a list of AlgebraOfGraphics.ProcessedLayer objects. ProcessedLayers objects are the output of the processing pipeline and can be drawn without further processing.

AlgebraOfGraphics.Entry — Type

Entry(plottype::PlotType, positional::Arguments, named::NamedArguments)

Define plottype as well as positional and named arguments for a single plot.

AlgebraOfGraphics.AxisEntries — Type

AxisEntries(axis::Union{Axis, Nothing}, entries::Vector{Entry}, categoricalscales, continuousscales)

Define all ingredients to make plots on an axis. Each categorical scale should be a CategoricalScale, and each continuous scale should be a ContinuousScale.