Data

AlgebraOfGraphics.data — Function

data(table)

Create a Layer with its data field set to a table-like object.

There are no type restrictions on this object, as long as it respects the Tables interface. In particular, any one of these formats should work out of the box.

To create a fully specified layer, the layer created with data needs to be multiplied with the output of mapping.

spec = data(...) * mapping(...)

using AlgebraOfGraphics
df = (a = rand(10), b = rand(10))
data(df)

Layer 
  transformation: identity
  data: AlgebraOfGraphics.Columns{@NamedTuple{a::Vector{Float64}, b::Vector{Float64}}}
  positional:
  named: