Drawing Layers

A Layer or Layers object can be plotted using the functions draw or draw!.

AlgebraOfGraphics.draw — Function

draw(s; axis=NamedTuple(), figure=NamedTuple, palettes=NamedTuple())

Draw a Layer or Layers object s. The output can be customized by giving axis attributes to axis, figure attributes to figure, or custom palettes to palettes. Legend and colorbar are drawn automatically. For finer control, use draw!, legend!, and colorbar! independently.

AlgebraOfGraphics.draw! — Function

draw!(fig, s; axis=NamedTuple(), palettes=NamedTuple())

Draw a Layer or Layers object s on fig. fig can be a figure, a position in a layout, or an axis if s has no facet specification. The output can be customized by giving axis attributes to axis or custom palettes to palettes.

Whereas draw automatically adds colorbar and legend, draw! does not, as it would be hard to infer a good default placement that works in all scenarios.

Colorbar and legend, should they be necessary, can be added separately with the colorbar! and legend! helper functions. See also Embedding AlgebraOfGraphics plots in a Makie figure for a complex example.

AlgebraOfGraphics.colorbar! — Function

colorbar!(figpos, grid; kwargs...)

Compute colorbar for grid (which should be the output of draw!) and draw it in position figpos. Attributes allowed in kwargs are the same as MakieLayout.Colorbar.

AlgebraOfGraphics.legend! — Function

legend!(figpos, grid; kwargs...)

Compute legend for grid (which should be the output of draw!) and draw it in position figpos. Attributes allowed in kwargs are the same as MakieLayout.Legend.